добавляет модули в проект
